Calibrate damage function parameters
Updated the README to reflect the process of calibrating the damage
function.  Specifically, the commit includes changes to: - Clarify the
source of the interest rate model parameters.  - Explain that the
preference parameters are chosen to minimize the discrepancy between
interest rate and per-capita income growth projections.  - Add a step
for saving the climate damage point estimates by Barrage and
Nordhaus (2024).  - Outline the process of choosing damage function
parameters by calibrating them against the Barrage and Nordhaus damage
point estimates.